+++ src/matchers/builtin/anything.js
@@ -1,0 +1,6 @@
+let create = require('../create')
+
+module.exports = create({
+  name: 'anything',
+  matches () { return true }
+})
